table: use generics throughout
Java 8 now gives no compile-time warnings.
Most of the changes are adding parameterisations to parameterised types,
in most cases the code is not changed otherwise in a way that will not
be erased at compile time, but in a few cases some minor changes
have been made to accommodate the generics.  Various other warnings
like redundant casts have been avoided too.
A lot of lines of code are changed here.

SuppressWarnings is used in a few places, notably when dealing with
classes like JComboBox that have acquired generics since java 6.

+++ table/src/main/uk/ac/starlink/table/join/EllipseToy.java
@@ -202,12 +202,12 @@ public class EllipseToy extends JComponent {
                                          e2_.y_ + Math.max( e2_.a_, e2_.b_ ) ));
                 Point p1 = Double.isNaN( match.alpha1_ + match.delta1_ )
                          ? null
-                         : new Point( (int) alphaToX( match.alpha1_, qp ),
-                                      (int) deltaToY( match.delta1_, qp ) );
+                         : new Point( alphaToX( match.alpha1_, qp ),
+                                      deltaToY( match.delta1_, qp ) );
                 Point p2 = Double.isNaN( match.alpha2_ + match.delta2_ )
                          ? null
-                         : new Point( (int) alphaToX( match.alpha2_, qp ),
-                                      (int) deltaToY( match.delta2_, qp ) );
+                         : new Point( alphaToX( match.alpha2_, qp ),
+                                      deltaToY( match.delta2_, qp ) );
                 if ( p1 != null ) {
                     g2.drawLine( e1_.x_, e1_.y_, p1.x, p1.y );
                 }
